Change datatype
In integer division, the digits after the decimal point is removed

+++ algorithm/P1008.java
@@ -7,8 +7,8 @@ public class P1008 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		
-		int A = sc.nextInt();
-		int B = sc.nextInt();
+		double A = sc.nextDouble();
+		double B = sc.nextDouble();
 		
 		if ( 0 < A && A < 10 && 0 <B && B < 10 ) {
 			System.out.println(A / B);
